Polish formatting and interactivity
All aria hints related to interactivity are now created by the JS.
After all, If the JS doesn't run, we won't have any interactivity.
So now, the hints are only present if they are relevant.

Also, add a button to expand/collapse all sections. Being able to scroll
through to get the ratings overview is nice, nice enough that I want it
to be the default. But asking the user to click 23 times in order to read
the details is a bit much.